+++ test/libsolidity/JSONCompiler.cpp
@@ -25,8 +25,8 @@
 #include <libsolidity/interface/Version.h>
 #include <libsolc/libsolc.h>
 
-#include "../Metadata.h"
-#include "../TestHelper.h"
+#include <test/Metadata.h>
+#include <test/Options.h>
 
 using namespace std;