+++ test/libsolidity/SemVerMatcher.cpp
@@ -25,7 +25,7 @@
 #include <tuple>
 #include <libsolidity/parsing/Scanner.h>
 #include <libsolidity/analysis/SemVerHandler.h>
-#include "../TestHelper.h"
+#include <test/Options.h>
 
 using namespace std;