+++ test/libsolidity/syntaxTests/docstring_empty_description.sol
@@ -1,0 +1,6 @@
+contract C {
+    /// @param id
+    function vote(uint id) public;
+}
+// ----
+// DocstringParsingError: No description given for param id